Setup bookdown project, add logo and math support
Since dissertations will of course be quite long documents, it will
likely be more convenient to have a modular document setup, for which
a bookdown project seems like the best way to go. So, I have now setup
the ability to start a new bookdown project for the dissertation in
addition to being able to do it in a standalone R Markdown document.

I also set up a logo and load the 'amsmath' and 'amssymb' packages in
the LaTeX template for math support.